+++ Games/Tools/world_editor/world_editor.py
@@ -21,6 +21,7 @@ BLACK = (0,0,0)
 class Main(Template):
     def __init__(self, size):
         Template.__init__(self, size)
+        self.name = "Under Construction"
         self.size = size
         self.block = 64
         self.dt = 0.
@@ -104,7 +105,7 @@ class Main(Template):
 
     def setup_loadmenu(self):
         folder = Path("./save")
-        filelist = [f for f in folder.files("*.sav")]
+        filelist = [f.basename() for f in folder.files("*.sav")]
         amount = len(filelist)
         menu = Menu((240, amount*21), (250, 21))
         menu.add_buttons(amount, filelist)
@@ -123,14 +124,11 @@ class Main(Template):
         size = (10, int(length/10))
         menu = Map(self.size, size, self.block, "tiles/Empty_tile_64p.png")
         menu.setup(CYAN)
-        i = 1
-        for tile in menu.group:
-            tile.filename = "maptile{}".format(i)
+        for i, tile in enumerate(menu.group):
+            tile.filename = "{}".format(i)
             tile.image = self.menu_list[i]
-            if i+1 >= len(self.menu_list):
+            if i+1 >= length:
                 break
-            else:
-                i += 1
         return menu
 
 #################### Key and Mouse Methods #####################################
@@ -255,13 +253,33 @@ class Main(Template):
             tile.dirty = 1
 
     def savemap(self):
+        data = {
+        "Name":self.name,
+        "Background":self.map1,
+        "Foreground":self.map2,
+        "Menu":self.menu
+        }
+        save = SaveMap(data)
+        save.write_to_file("./save/savefile.sav")
+
+    def loadmap(self):
+        loadmap = LoadMap()
+        data = loadmap.load_from_file("./save/savefile.sav")
+        self.name = data["Name"]
+        self.map1 = data["Background"]
+        self.map2 = data["Foreground"]
+        self.menu = data["Menu"]
+        print("map loaded")
+
+
+    def old_savemap(self):
         size = tuple_mult((40,40), self.block)
         save = SaveMap("savemap", size, self.block, "img/magecity_64p.png",
                         self.map1.group, self.map2.group)
         save.write_to_file()
         self.floating_text.set_text("Map Saved", True)
 
-    def loadmap(self):
+    def old_loadmap(self):
         select = None
         for button in self.load_menu.buttons:
             if button.clicked:
@@ -278,7 +296,8 @@ class Map(object):
         self.screen_size = screen_size
         self.grid = grid
         self.block = block
-        self.image = image
+        self.image_string = image
+        self.image = pg.image.load(image).convert_alpha()
         self.size = (self.grid[0]*self.block, self.grid[1]*self.block)
         self.rect = pg.Rect(0,0,screen_size[0]/2, screen_size[1]/2)
         self.map = pg.Surface(self.size)
@@ -291,6 +310,7 @@ class Map(object):
         self.clipped = False
         self.selected = None
         self.sel_old = None
+        self.saved_surf = None
 
     def setup(self, color, alpha=255):
         self.alpha = alpha
@@ -303,7 +323,10 @@ class Map(object):
         group = pg.sprite.LayeredDirty()
         for y in range(self.grid[1]):
             for x in range(self.grid[0]):
-                tile = Tile(self.image, (x*self.block, y*self.block))
+                tile = Tile(None, xy=(x*self.block, y*self.block))
+                tile.image = self.image
+                tile.rect = tile.image.get_rect()
+                tile.rect.topleft = tile.xy
                 tile.dirty = 1
                 group.add(tile)
         return group
